Modularize enums
