+++ openfisca_core/indexed_enums/config.py
@@ -1,0 +1,3 @@
+import numpy
+
+ENUM_ARRAY_DTYPE = numpy.int16
